[refactor] Move common code to utils and reuse some functions
